- change float to double - add checking of results correction by sqrt norm - add pretty console output - add calculations of GPU peac performance

+++ kernels/naive_mult.cl
@@ -1,13 +1,13 @@
-__kernel void multiply(const __global float* A,
-                       const __global float* B,
-                       __global float* C,
+__kernel void multiply(const __global double* A,
+                       const __global double* B,
+                       __global double* C,
                        const int widthA, 
                        const int widthB) 
 {
     const int x = get_global_id(0);
     const int y = get_global_id(1);
  
-    float result = 0.0;
+    double result = 0.0;
     for (int i = 0; i < widthA; ++i) 
     {
         result += A[widthA * y + i] * B[widthB * i + x];
